refactor: deduplicate sanitizeFileMode, escapeNonDerivedVars, copyFile and remove dead code
- Extract SanitizeFileMode and TextSampleSize to fileutil package, replacing
  duplicate implementations in scaffold/output.go and templateupdate/renderer.go
- Consolidate escapeNonDerivedVars into standalone function in scaffold/processor.go,
  eliminating identical method bodies on DefaultPathProcessor and DefaultOutputWriter
- Replace local copyFile in templateupdate/backup.go with fileutil.CopyFile (adds
  fsync and setuid bit stripping)
- Remove unused WithEngine/WithValidator scaffold options, MockHookRunner test mock,
  and NewLoaderFS test helper (all flagged by deadcode)
- Move TestUT_SanitizeFileMode to fileutil package where the function now lives
